Add auto importing for route modules
