+++ Backend/server/routes/user.py
@@ -1,7 +1,5 @@
 from . import Router
 
-__all__ = ['user_router']
-
 user_router = Router(prefix='/user', tags=['user'])
